Delete query results after a package is modified in some way
Move PackageCommitEnumerator around

+++ ExplorePackages/Entities/PackageQueryMatch.cs
@@ -2,7 +2,7 @@
 {
     public class PackageQueryMatch
     {
-        public int Key { get; set; }
+        public long Key { get; set; }
         public int PackageKey { get; set; }
         public int PackageQueryKey { get; set; }
         public Package Package { get; set; }
